requirements.txt
irc==18.0.0
python-dotenv==0.10.5
requests==2.22.0
